requirements.txt
ccxt>=2.9.0
pandas>=1.5.0
numpy>=1.22.0
aiohttp>=3.8.0
asyncio>=3.4.3
pydantic>=1.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.3.0
pyyaml>=6.0.0
prometheus-client>=0.15.0
websockets>=10.3
orjson>=3.8.0
python-telegram-bot>=13.13
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.0.0
mypy>=0.931
pylint>=2.13.0
